backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
psycopg2-binary==2.9.5
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==9.4.0
transformers==4.27.1
torch==2.0.0
cohere==4.5.1
instabot==0.117.0
# No scheduler required
gunicorn==20.1.0
psutil==5.9.5
PyJWT==2.6.0